feat: ajustes para criar o livro database
